[#9] refactor: FormValues interface 수정
