--- src/models/auth.ts
@@ -1,3 +1,0 @@
-export interface FormValues {
-  [key: string]: string
-}
