+++ src/models/signIn.ts
@@ -1,0 +1,4 @@
+export interface FormValues {
+  email: string
+  password: string
+}
